douleに変換して返すよう修正

+++ cpp06/ex00/Double_Format.cpp
@@ -7,7 +7,7 @@ double convertDouble(std::string& val) {
   if (errno == ERANGE) {
     throw ImpossibleException();
   }
-  return static_cast<float>(res);
+  return static_cast<double>(res);
 }
 
 void printDoubleFormat(std::string& val) {
